test(auth): move project key coverage to v2
